Game class refactor

+++ absract_game.cpp
@@ -1,6 +1,6 @@
 #include "absract_game.h"
 
-AbstractGame::AbstractGame(unsigned int window_width, unsigned int window_height,
+AbstractGame::AbstractGame(float window_width, float window_height,
                            sf::String title, sf::Time time_per_frame)
         :window(sf::VideoMode(window_width, window_height), title), TimePerFrame(time_per_frame) { }
